# Reduce performance regression from on-CPU mapping of labels before rendering