… napari/benchmarks/benchmark_labels_layer.py
… napari/benchmarks/utils.py
… napari/utils/colormaps/_tests/test_colormap.py
… napari/utils/colormaps/_tests/test_colormap_utils.py
… napari/utils/colormaps/colormap.py