# References and relevant issues
closes #6579
closes #6578

# Description

This PR fixes two problems. Incorrect painting in direct mode (by fix setup cache). 
Increases performance of drawing by cache mapping from label to transformed values and using it during paint. It also requires of changes in cache. Previously it was filled with zeros that require to perform mapping on the whole data slice. When update pixels using only cached mapping, it leads to artifact like in this screenshot. (from https://github.com/napari/napari/issues/6579#issuecomment-1882913886)

![image](https://github.com/napari/napari/assets/3826210/2e8b514a-7574-4e45-869d-164323f1e92c)

So the current cache is initialized with raw and view values. 

PR also contains fixes in benchmark to ensure that data are not shared between benchmarks (what happens because of `lru_cache` usage).