@jni 2024-01-11
@ksofiyuk I'm curious whether in this current state the painting is fast enough that you would consider #6579 closed? We can always bring back the `_all_vals` caching from pre-#3308, but I want to avoid it unless necessary, as it was quite finicky e.g. because of negative indexing.

I'm also curious whether using uint8/uint16 labels data completely removes the bottleneck for you.

At any rate @Czaki I think the mapping of single values while painting is a nice idea, thanks for spotting it so quickly!