@ksofiyuk 2024-01-11
I've been testing this PR, the performance seems to be improved, however, I think using numba here is a mistake, especially for the modulo operation, it only hurts the performance.

I've used the following code to benchmark it:
```python
n = 50
for msize in [2 ** x for x in range(6, 11)]:
    values = np.random.randint(100000, size=(msize, msize), dtype=np.int32)
    
    print(f"{values.size=}")
    
    start_time = time.time()
    for i in range(1000):
        _zero_preserving_modulo_numpy(values, n=n, dtype=np.uint8)
    print(f'_zero_preserving_modulo_numpy, elapsed time: {time.time()-start_time:.3f} s')
    
    _zero_preserving_modulo(values, n=n, dtype=np.uint8) # warmup numba
    start_time = time.time()
    for i in range(1000):
        _zero_preserving_modulo(values, n=n, dtype=np.uint8)
    print(f'_zero_preserving_modulo, elapsed time: {time.time()-start_time:.3f} s')
    print()
```

Here are results:
```
values.size=4096
_zero_preserving_modulo_numpy, elapsed time: 0.036 s
_zero_preserving_modulo, elapsed time: 0.865 s

values.size=16384
_zero_preserving_modulo_numpy, elapsed time: 0.100 s
_zero_preserving_modulo, elapsed time: 1.365 s

values.size=65536
_zero_preserving_modulo_numpy, elapsed time: 0.224 s
_zero_preserving_modulo, elapsed time: 0.530 s

values.size=262144
_zero_preserving_modulo_numpy, elapsed time: 0.938 s
_zero_preserving_modulo, elapsed time: 0.867 s

values.size=1048576
_zero_preserving_modulo_numpy, elapsed time: 4.531 s
_zero_preserving_modulo, elapsed time: 2.230 s
```

First, it starts to show benefits only from very large data size (closer to 1000000, not to just 1000). Second, it is actually much slower per CPU core even for `values.size=1048576`. numba uses **all the 16 cores of my CPU** and it is only twice faster than a single-core numpy code. Third, it can introduce random performance drops because of spawning threads for small data sizes.