@ksofiyuk 2024-01-11
And because of numba the UI makes less responsive at the beginning. At the very beginning of the drawing process the UI can freeze for 1-2 sec because of the JIT compilation. It also freezes for a second or so when I switch between auto and direct mode for the first time.