@Czaki 2024-01-11
> At the very beginning of the drawing process the UI can freeze for 1-2 sec because of the JIT compilation

No. This freeze comes from memory allocation for the paint cache. In exception to your PR, the cache is allocated only when someone starts drawing to save memory. 

> numba uses **all the 16 cores of my CPU** and it is only twice faster than the single-core numpy code.

And numba uses ~3 times less memory, as do not need to use additional arrays. 

Yes. The better solution will be to use compiled code (ex. cython) but it is not currently available. 

And your perspective is to work with big, single-plane images. But napari need to be performant also for 3d data. 

Please provide your system configuration. on my macbook m2 it is:

```
values.size=4096
_zero_preserving_modulo_numpy, elapsed time: 0.019 s
_zero_preserving_modulo, elapsed time: 0.079 s

values.size=16384
_zero_preserving_modulo_numpy, elapsed time: 0.034 s
_zero_preserving_modulo, elapsed time: 0.111 s

values.size=65536
_zero_preserving_modulo_numpy, elapsed time: 0.121 s
_zero_preserving_modulo, elapsed time: 0.085 s

values.size=262144
_zero_preserving_modulo_numpy, elapsed time: 0.502 s
_zero_preserving_modulo, elapsed time: 0.141 s

values.size=1048576
_zero_preserving_modulo_numpy, elapsed time: 2.089 s
_zero_preserving_modulo, elapsed time: 0.372 s
```

And when I use it for 3d data (added 40 layers it your code by changing `(msize, msize)` to `(40, msize, msize)`

the results are:

```
values.size=163840
_zero_preserving_modulo_numpy, elapsed time: 0.307 s
_zero_preserving_modulo, elapsed time: 0.108 s

values.size=655360
_zero_preserving_modulo_numpy, elapsed time: 1.407 s
_zero_preserving_modulo, elapsed time: 0.264 s

values.size=2621440
_zero_preserving_modulo_numpy, elapsed time: 5.210 s
_zero_preserving_modulo, elapsed time: 0.607 s

values.size=10485760
_zero_preserving_modulo_numpy, elapsed time: 25.374 s
_zero_preserving_modulo, elapsed time: 2.350 s

values.size=41943040
_zero_preserving_modulo_numpy, elapsed time: 114.739 s
_zero_preserving_modulo, elapsed time: 7.802 s
```


But from this code, it looks like we may raise the threshold from 1000 to 50000. @jni What did you think?