@ksofiyuk 2024-01-11
@Czaki Even on your laptop, even on very large arrays the tests do not justify using numba. Your results show that it takes 25 seconds to process a very large array of `10485760` elements. However, keep in mind, that it is time for 1000 loops. So, in reality it will take only 25 ms of a single core time to process it. 

> No. This freeze comes from memory allocation for the paint cache. In exception to your PR, the cache is allocated only when someone starts drawing to save memory.

The part about the memory allocation is invalid explanation for freezes because when I disable numba functions I do not experience freezes. Moreover, with numba freezes happen even on very small images.

By the way, did you check it? I've checked and the very first call of this numba function takes 500-600 ms on my machine even for a very small matrix.

```python
values = np.random.randint(100000, size=(8, 8), dtype=np.int32)

start_time = time.time()
_zero_preserving_modulo(values, n=50, dtype=np.uint8)
print(f"first call time: {time.time() - start_time:.3f} s")

start_time = time.time()
_zero_preserving_modulo(values, n=50, dtype=np.uint8)
print(f"second call time: {time.time() - start_time:.3f} s")
```

```
first call time: 0.536 s
second call time: 0.054 s
```

The same story for `_labels_raw_to_texture_direct`. I've been testing it also, the very first call takes **1.72 s**! And exactly this JIT complication causes the freeze at the beginning when you draw in the direct mode.