@Czaki 2024-01-11
> The part about the memory allocation is invalid explanation for freezes because when I disable numba functions I do not experience freezes. Moreover, with numba freezes happen even on very small images.

numba is also used for the initial mopping of the image before sending to GPU. So when you start painting it is not the first call of numba. During implementation, I observed the same problem when move the initialization of cache. Please remove conditional setup of cache in `raw_to_displayed` and check then. 

> I've checked and the very first call of this numba function takes 500-600 ms on my machine even for a very small matrix.

Yes. This is compilation time.