@Czaki 2024-01-13
current benchmarks report 2 times worse performance for drawing:

```

| Change   | Before [4a1b0b54]    | After [b096b49c]    |   Ratio | Benchmark (Parameter)                                                                          |
|----------|----------------------|---------------------|---------|------------------------------------------------------------------------------------------------|
| +        | 34.3±0.1μs           | 120±0.3μs           |    3.5  | benchmark_labels_layer.Labels3DSuite.time_raw_to_displayed(32, <class 'numpy.uint32'>)         |
| +        | 26.5±0.09μs          | 73.2±0.9μs          |    2.76 | benchmark_labels_layer.Labels2DSuite.time_raw_to_displayed(128, <class 'numpy.int32'>)         |
| +        | 58.7±0.8μs           | 119±0.7μs           |    2.03 | benchmark_labels_layer.Labels2DColorDirectSuite.time_paint_circle(2048, <class 'numpy.uint8'>) |
| +        | 62.1±0.6μs           | 122±1μs             |    1.96 | benchmark_labels_layer.Labels2DColorDirectSuite.time_paint_circle(1024, <class 'numpy.uint8'>) |
| +        | 57.0±0.8μs           | 111±0.4μs           |    1.96 | benchmark_labels_layer.Labels2DColorDirectSuite.time_paint_circle(4096, <class 'numpy.uint8'>) |
| +        | 65.3±0.5μs           | 124±0.3μs           |    1.91 | benchmark_labels_layer.Labels2DColorDirectSuite.time_paint_circle(32, <class 'numpy.uint8'>)   |
| +        | 64.8±0.4μs           | 123±0.5μs           |    1.9  | benchmark_labels_layer.Labels2DColorDirectSuite.time_paint_circle(256, <class 'numpy.uint8'>)  |
| +        | 64.6±0.4μs           | 123±0.9μs           |    1.9  | benchmark_labels_layer.Labels2DColorDirectSuite.time_paint_circle(512, <class 'numpy.uint8'>)  |
| +        | 65.3±0.2μs           | 123±0.6μs           |    1.89 | benchmark_labels_layer.Labels2DColorDirectSuite.time_paint_circle(16, <class 'numpy.uint8'>)   |
| +        | 65.2±0.3μs           | 123±0.3μs           |    1.89 | benchmark_labels_layer.Labels2DColorDirectSuite.time_paint_circle(64, <class 'numpy.uint8'>)   |
| +        | 65.5±0.2μs           | 123±0.9μs           |    1.88 | benchmark_labels_layer.Labels2DColorDirectSuite.time_paint_circle(128, <class 'numpy.uint8'>)  |
| +        | 20.8±0.1μs           | 33.5±0.09μs         |    1.61 | benchmark_labels_layer.Labels2DSuite.time_raw_to_displayed(64, <class 'numpy.int32'>)          |
| +        | 23.1±0.1μs           | 37.3±0.2μs          |    1.61 | benchmark_labels_layer.Labels3DSuite.time_raw_to_displayed(16, <class 'numpy.uint32'>)         |
```

I need to investigate why.