@Czaki 2024-01-16
Current benchmark status. 

```

| Change   | Before [4a1b0b54]    | After [036aebd4]    |   Ratio | Benchmark (Parameter)                                                                                     |
|----------|----------------------|---------------------|---------|-----------------------------------------------------------------------------------------------------------|
| -        | 4.10±0.01ms          | 2.71±0.01ms         |    0.66 | benchmark_labels_layer.Labels2DColorDirectSuite.time_refresh(64, <class 'numpy.int32'>)                   |
| -        | 4.09±0.02ms          | 2.65±0.03ms         |    0.65 | benchmark_labels_layer.Labels2DColorDirectSuite.time_refresh(32, <class 'numpy.int32'>)                   |
| -        | 4.07±0.02ms          | 2.60±0.01ms         |    0.64 | benchmark_labels_layer.Labels2DColorDirectSuite.time_refresh(16, <class 'numpy.int32'>)                   |
| -        | 3.81±0.02ms          | 2.42±0.03ms         |    0.64 | benchmark_labels_layer.Labels2DColorDirectSuite.time_update_thumbnail(128, <class 'numpy.int32'>)         |
| -        | 3.90±0.03ms          | 2.50±0.02ms         |    0.64 | benchmark_labels_layer.Labels2DColorDirectSuite.time_update_thumbnail(4096, <class 'numpy.int32'>)        |
| -        | 3.84±0.01ms          | 2.41±0.01ms         |    0.63 | benchmark_labels_layer.Labels2DColorDirectSuite.time_update_thumbnail(16, <class 'numpy.int32'>)          |
| -        | 3.83±0.01ms          | 2.40±0.01ms         |    0.63 | benchmark_labels_layer.Labels2DColorDirectSuite.time_update_thumbnail(32, <class 'numpy.int32'>)          |
| -        | 3.82±0.01ms          | 2.40±0.01ms         |    0.63 | benchmark_labels_layer.Labels2DColorDirectSuite.time_update_thumbnail(64, <class 'numpy.int32'>)          |
| -        | 18.6±0.2μs           | 11.6±0.07μs         |    0.62 | benchmark_labels_layer.Labels2DSuite.time_raw_to_displayed_full(32, <class 'numpy.int32'>)                |
| -        | 18.4±0.09μs          | 11.0±0.02μs         |    0.6  | benchmark_labels_layer.Labels2DSuite.time_raw_to_displayed_full(16, <class 'numpy.int32'>)                |
| -        | 10.7±0.06μs          | 5.72±0.04μs         |    0.54 | benchmark_labels_layer.Labels2DSuite.time_raw_to_displayed_sliced(4096, <class 'numpy.uint8'>)            |
| -        | 10.4±0.06μs          | 5.47±0.02μs         |    0.53 | benchmark_labels_layer.Labels2DSuite.time_raw_to_displayed_sliced(128, <class 'numpy.uint8'>)             |
| -        | 10.4±0.06μs          | 5.50±0.04μs         |    0.53 | benchmark_labels_layer.Labels2DSuite.time_raw_to_displayed_sliced(16, <class 'numpy.uint8'>)              |
| -        | 10.5±0.1μs           | 5.61±0.03μs         |    0.53 | benchmark_labels_layer.Labels2DSuite.time_raw_to_displayed_sliced(2048, <class 'numpy.uint8'>)            |
| -        | 10.3±0.03μs          | 5.52±0.02μs         |    0.53 | benchmark_labels_layer.Labels2DSuite.time_raw_to_displayed_sliced(256, <class 'numpy.uint8'>)             |
| -        | 10.3±0.09μs          | 5.50±0.04μs         |    0.53 | benchmark_labels_layer.Labels2DSuite.time_raw_to_displayed_sliced(32, <class 'numpy.uint8'>)              |
| -        | 10.6±0.04μs          | 5.62±0.02μs         |    0.53 | benchmark_labels_layer.Labels2DSuite.time_raw_to_displayed_sliced(512, <class 'numpy.uint8'>)             |
| -        | 10.4±0.02μs          | 5.49±0.01μs         |    0.53 | benchmark_labels_layer.Labels2DSuite.time_raw_to_displayed_sliced(64, <class 'numpy.uint8'>)              |
| -        | 10.3±0.05μs          | 5.35±0.08μs         |    0.52 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced(4096, <class 'numpy.uint8'>) |
| -        | 10.7±0.2μs           | 5.61±0.03μs         |    0.52 | benchmark_labels_layer.Labels2DSuite.time_raw_to_displayed_sliced(1024, <class 'numpy.uint8'>)            |
| -        | 10.8±0.1μs           | 5.27±0.05μs         |    0.49 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced(2048, <class 'numpy.uint8'>) |
| -        | 12.2±0.09μs          | 5.28±0.03μs         |    0.43 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced(1024, <class 'numpy.uint8'>) |
| -        | 12.2±0.07μs          | 5.12±0.02μs         |    0.42 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced(128, <class 'numpy.uint8'>)  |
| -        | 12.1±0.1μs           | 5.09±0.02μs         |    0.42 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced(16, <class 'numpy.uint8'>)   |
| -        | 12.1±0.05μs          | 5.14±0.01μs         |    0.42 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced(256, <class 'numpy.uint8'>)  |
| -        | 12.2±0.07μs          | 5.14±0.03μs         |    0.42 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced(32, <class 'numpy.uint8'>)   |
| -        | 12.4±0.2μs           | 5.25±0.08μs         |    0.42 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced(512, <class 'numpy.uint8'>)  |
| -        | 12.2±0.07μs          | 5.17±0.03μs         |    0.42 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced(64, <class 'numpy.uint8'>)   |
| Change   | Before [4a1b0b54]    | After [036aebd4]    |   Ratio | Benchmark (Parameter)                                                                                         |
|----------|----------------------|---------------------|---------|---------------------------------------------------------------------------------------------------------------|
| +        | 28.5±0.2μs           | 264±4μs             |    9.28 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced_num(4096, <class 'numpy.int32'>) |
| +        | 27.9±0.09μs          | 228±5μs             |    8.15 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced_num(2048, <class 'numpy.int32'>) |
| +        | 27.6±0.1μs           | 121±7μs             |    4.39 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced_num(1024, <class 'numpy.int32'>) |
| +        | 46.0±0.4μs           | 192±5μs             |    4.17 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced(4096, <class 'numpy.int32'>)     |
| +        | 28.1±0.6μs           | 115±6μs             |    4.09 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced_num(512, <class 'numpy.int32'>)  |
| +        | 27.9±0.3μs           | 110±5μs             |    3.95 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced_num(256, <class 'numpy.int32'>)  |
| +        | 44.9±0.7μs           | 170±10μs            |    3.79 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced(2048, <class 'numpy.int32'>)     |
| +        | 27.1±0.2μs           | 90.0±4μs            |    3.32 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced_num(32, <class 'numpy.int32'>)   |
| +        | 219±0.3μs            | 452±6μs             |    2.07 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_full(128, <class 'numpy.int32'>)        |
| +        | 44.3±0.4μs           | 77.1±3μs            |    1.74 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced(1024, <class 'numpy.int32'>)     |
| +        | 45.4±0.5μs           | 75.1±5μs            |    1.65 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced(512, <class 'numpy.int32'>)      |
| +        | 82.1±0.3μs           | 134±3μs             |    1.63 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_full(64, <class 'numpy.int32'>)         |
| +        | 45.1±0.5μs           | 68.9±3μs            |    1.53 | benchmark_labels_layer.Labels2DColorDirectSuite.time_raw_to_displayed_sliced(256, <class 'numpy.int32'>)      |
| +        | 380±5μs              | 582±7μs             |    1.53 | benchmark_labels_layer.Labels2DColorDirectSuite.time_set_view_slice(128, <class 'numpy.int32'>)               |
```

Based on my checks, the source of the problem may be numba compilation time. I'm working on an update of setup function.